requirements.txt
google-api-python-client
python-dotenv
textblob
pandas
googletrans
